Change website to follow Odin requirements
I added everything to the homepage instead of the recipe page so
I moved the code around so that everything is in the right place.

I also played around with adding a link to the list on the homepage
which took a few attemtps as I haven't watched a tutorial about
putting lots of tags together. The same applies to when I added a
bold and underlined title